requirements.txt
seqeval==0.0.12
tqdm==4.30.0
ipywidgets==7.4.2
Keras==2.3.1
transformers==2.8.0
joblib==0.13.2
torch==1.2.0
tensorflow==1.15.2
numpy==1.18.1
sphinx_rtd_theme==0.4.3
pandas==1.0.1
scikit_learn==0.23.1
PyYAML==5.3.1
